Make basic template easier to clean up (#12217)
